Find images open in multiple documents, user select only one's settings
add store/forget buttons to exclusive (but uncheckable) button group
for a file when multiple documents with that filename are open.

user can still export from any of those documents individually but only
the settings from one will be saved and persist to next run of dialog.

also inform user on statusbar. Might need to work on the wording later
as it seems to suggest to user that there's a problem when really it's
just an fyi.