Prompt user to re-use settings from previous versions of image if found
on pressing export for an image with no settings stored, look to see if
there are stored settings for a file that could be a previous version
of this image (eg. 'my_image.kra' from 'my_image_001.kra').

if yes, ask user if they want to:
- make a copy of old image settings for new image;
- convert the old image settings to the new image (replace old path);
- ignore old versions, open config dialog like when none exist.
- do nothing

copy/replace will immediately update settings and export the image.

future improvements might include capturing an incremental save as it
happens and automatically replacing previous version settings (or not
as per user preferences); and checking for files older than one version
ago.