Add option for behaviour when newly exporting unstored versioned images
previous behaviour was to always ask. now user can choose to auto-
matically copy, replace or make new settings, or to still ask.

new default behaviour is automatically replace as that makes the most
sense 99% of the time.

note that if multiple candidates for previous versions are found (for
some reason), then the prompt must still appear.