Add options for custom icons and respond to light/dark theme changes
by default, uses the svg images in the plugin folder as icons.
disabling custom icons will fall back to Krita default icons as before.

tracks the current theme by looking at the check state of the menu item
actions in the Settings/Themes menu. for themes shipped with Krita, the
correct icons to use are known. for additional themes, the theme name
is used to guess (eg. is 'dark' in the name). if it all goes wrong the
user has the option to just force use of light or dark icons anyway.