Hide settings widgets for export types other than selected
widgets are first hidden after setup is finished and columns have been
sized so as not to break column sizing.

note: this isn't a very scalable approach and there's no elegant way to
hide columns based on export types (columns need to show for all types
being exported, and switching visible columns based on type of last-
interacted-with file or with radio buttons will be intrusive and
annoying).
I expect to ditch this column-per-setting approach and have a single
column with swappable groups of settings widgets inside.