Add ability to set export scale
scale settings can be adjusted, and whether to scale can be turned on
and off separately. adjustment is disabled for unopened documents as
the size and resolution need to be available.

a document can not be set with different scale settings per export
image type (unlike eg. metadata and author data).

issue: document metadata is absent from exported image if image was
scaled at export. I see the same thing happen with the native export
advanced path, so I think this is a krita bug.

notes on scale dialog:

spinboxes allow only for basic number entry and use of % to set to
percentage of initial value (eg. type '%' at end of '50' to set spinbox
that initialised as 1380 to 690 - 50% of 1380). does not accept
equations. also units can not be changed: the pixel dimensions are in
pixels and the print dimensions are in inches.

the print dimensions part of this dialog has been cobbled together by
tweaking its behaviour until its output corresponded with that of the
native scale dialog. it might still deviate though, and at some point
it should probably be rewritten into a straight python copy of the
native dialog code instead. (ideally the dialog would be accessible
from the plugin api.)

I think 99% of the time people aren't exporting to print, and 99% of
those times they are they should have no issues here. if needs be they
can still make one-off use of the normal export path to ensure correct-
ness, so further work on this can safely wait until after other higher-
priority stuff.

the warning about image size is also a bit jank, but I think it's
better to have the heads-up than to not. the memory usage estimate
accuracy is unverified.