Multi-line elided text utility module (scripter test code)
script can be run standalone in krita scripter.

flows text into a rectangle of a given width up to a maximum allowed
number of lines. if the text is too long (2), will cut out the middle
of the text and replace with ellipses (3).

to do multi-line elide, the text is flowed forward from the start up to
the elide point, and backwards from the end to the elide point.

an earlier version simply flowed the full text as in (2) then displaced
the trailing lines up to the elide, removing the intervening overflow
lines. (4) shows typical results: the last line has a lot of wasted
space at the end. flowing the text backwards (3) wastes less space.

┌───────────(1)─┐     ┌───────(2)─┐     ┌───────(3)─┐     ┌───────(4)─┐
│the quick brown│     │the quick  │     │the quick  │     │the quick  │
│fox jumped over│     │brown fox  │     │brow … the │     │brow … lazy│
│the lazy dog   │ --- │jumped over│ --- │lazy dog   │ --- │dog        │
└───────────────┘     │the lazy   │     └───────────┘     └───────────┘
                      │dog        │
                      └───────────┘
full text fits        too long, if      text elided       naive elide,
in max lines          not elided        to fit            no rev-reflow

limitations
- has some edge cases when the rect width is about a character's width
  or less, which is worked around by just not displaying anything at or
  below an 'M' width.
- only spaces and forward slashes are considered as break points for
  wrapping, otherwise the backwards text flow would wrap punctuation
  marks to the wrong place eg. "end|. start" instead of "end.| start".
  the fix might be trivial but I've not tried.
- not tested with arabic chars, ligatures, any weird control chars or
  multi-code chars, emoji's, newlines, etc.
- could probably be optimised a lot (besides not being in python).