Allow user to select which extensions to show in the file type combobox
the previous commit filled the file type combobox lists with mostly
unwanted clutter. this commit adds a menu to the dialog options menu to
let the user pick exactly which file types to show.

the default is to show .png, .jpg and .jpeg, as it was previously. if
only one type is selected, the user can't deselect it without first
selecting another. if the saved set is invalid the default set is re-
instated.

if an image is set to export as a type that is hidden in the file type
combobox, the chosen export type will not be changed unless and until
the user selects another type, at which point they will have to re-show
the old type if they wish to change it back.

(I did try to use a QSortFilterProxyModel to do the filtering, but I
couldn't get it to not change the chosen export type if the type was
hidden.)