Allow exporting to any directory, add configuration controls
can now export to any directory, and its path can be stored as either
absolute or relative. the path, file name and extension can be set
together with a file dialog (which also asks user if a file should be
overwritten); the name and extension can still be set separately.

absolute export paths are saved beginning with "/", relative paths
with "./", and files in the same directory are saved without any
prefix (so backwards compatible with existing saved settings).

currently this is a distinction with little meaning but I expect to
add some way of dealing with moved source files at some point.