Add slider for setting dialog alternating row colour contrast
with the different row highlight colours, throwing alternating row
colours on top can hinder readability instead of help. now the user can
turn it off altogether or set it to just barely visible.

mixing in rgb is frowned upon because they usually produce unexpected
colours in the middle of the mix, but here the base and alternate
colours should be very close to begin with so there's just not enough
room for this to be an issue.

also the rgb values are slightly offset from each other before rounding
to integers to increase the apparent precision, eg. instead of jumping
from 5-5-5 to 6-6-6, will also step through 5-5-6 and 5-6-6 in between.