Refactor dialog row height
back when I started this I didn't know how to make this work so kept
adding kludges until it did work, then didn't want to try and unpick
the brittle code to strip it down to what was actually required. this
commit does that. it should all generally be nicer to work with now.

also, minor note: the Qt docs say that scheduleDelayedItemsLayout will
only re-layout once at event processing even if called multiple times
before. that may be true, but for whatever reason, wrapping the call in
a single-shot zero-interval qtimer greatly improved performance when
resizing columns.