Warn user if exporting the file would overwrite one of its source files
eg. export settings of "path/to/pic.jpg" will export "path/to/pic.jpg".

this will warn user if overwriting any existing version of its source
image; it won't warn for names that might potentially conflict with as-
yet unsaved versions, or for names that overwrite other files for which
settings are stored. this could be improved in the future.

as with names with multiple extensions, this doesn't prevent the user
from exporting in this state.

(previously commited but clobbered by a git reset, whoops.)