MultilineElidedText: fix edge case when eliding leading to crash
when max lines is 2 so the ellipsis is on line 2, and the truncation of
the word after the ellipsis to make room for it removes all the text
after the ellipsis, would try to set first_word to index out of range
in wrapped_text_b.

also, don't show ellipsis on last line if all text removed on it.

also fix number typo - loops overrun at 9999, not 100.