Fix unversioned file wrongly excluded as base of versioned files
"all" and "all_forward" versioning should behave the same for
unversioned files.

eg. for file "pic_002.kra", "pic.kra" with "all" mode should be recog-
nised as the base version, but was wrongly discarded. this is a special
case where it should be included.

considering the following examples cases:

1. "pic.kra" (all)       -> "pic_001.kra", "pic_002.kra", etc.
2. "pic.kra" (a_fwd)     -> "pic_001.kra", "pic_002.kra", etc.
3. "pic_001.kra" (all)   -> "pic_001_001.kra", "pic_001_002.kra", etc.
4. "pic_001.kra" (a_fwd) -> "pic_002.kra", "pic_003.kra", etc.

case 1 was being treated like case 3.