QETree: highlight row applying to doc, re-enable exporting unstored
previously highlighted/focused the item with the source path matching
the document, even though it may actually be exported with the settings
of another item (its base, if it's a version of it). now highlights the
base instead.

also recent changes made it so you couldn't export unstored images any-
more. this was unintentional but I considered keeping it because it
made it less ambiguous as to which item held the settings applying to
the document. in the end decided it should be allowed because otherwise
the user will be required to turn on advanced mode or pointlessly
twiddle a setting in order to enable the export button for a new doc.