Dialog: add file name filter field
only show items whose source file name starts with the text in the
file name filter field, if any. user can double click or start typing
to show suggestions in a popup. hard part was getting the popup to show
when the field is empty, and be wide enough to show the longest names
(the qt default makes it the width of the lineedit widget).