Dialog: add folder filter button
will only show items belonging to folders that are checked in the tree
in the folder filter button popup. if none checked, shows all.

the user can click on a folder to check it. when a folder is checked
all of its unchecked descendents become partially-checked. when a
folder is unchecked, all of its partially-checked descendents become
unchecked. this cascades down from the clicked folder but stops
wherever it meets another checked folder.

the user can also select rows of the tree to check or uncheck directly.

for each item in the dialog, its source file directory will be added
to the popup if it isn't there already. for each folder, all of its
ancestors back to the file system root will also be added. for example,
"/home/user/Pictures/pic.kra" will add / -> home -> user -> Pictures to
the popup tree. the result is a sort of sparse view of the file system.

new folders are added to the popup tree each time the dialog opens.
previously added folders that are no longer used are not currently
removed. filter settings do not persist between Krita sessions.